#29 Refactor command controllers folder
