+++ src/commands/index.js
@@ -1,12 +1,12 @@
 const ABaseCommand = require('./ABaseCommand')
 const CommandError = require('./CommandError')
 const commandList = require('./commandList')
-const CreateWalletCommand = require('./CreateWalletCommand')
-const DepositEthCommand = require('./DepositEthCommand')
-const DepositTokenCommand = require('./DepositTokenCommand')
-const GetAddressCommand = require('./GetAddressCommand')
-const GetBalanceCommand = require('./GetBalanceCommand')
-const OrderSignCommand = require('./OrderSignCommand')
+const CreateWalletCommand = require('./trading-wallet/CreateWalletCommand')
+const DepositEthCommand = require('./trading-wallet/DepositEthCommand')
+const DepositTokenCommand = require('./trading-wallet/DepositTokenCommand')
+const GetAddressCommand = require('./trading-wallet/GetAddressCommand')
+const GetBalanceCommand = require('./trading-wallet/GetBalanceCommand')
+const OrderSignCommand = require('./order/OrderSignCommand')
 
 module.exports = {
   ABaseCommand,
