+++ src/commands/order/OrderCancelCommand.js
@@ -1,5 +1,5 @@
-const ABaseCommand = require('./ABaseCommand')
-const CommandArg = require('../models/CommandArg')
+const ABaseCommand = require('../ABaseCommand')
+const CommandArg = require('../../models/CommandArg')
 
 /**
  * Class representing OrderCancelCommand. */
