+++ src/commands/token/GetAllowanceCommand.js
@@ -1,8 +1,8 @@
 const _ = require('lodash')
 
-const ABaseCommand = require('./ABaseCommand')
-const CommandArg = require('../models/CommandArg')
-const { Erc20TokenServiceBuilder } = require('../factories')
+const ABaseCommand = require('../ABaseCommand')
+const CommandArg = require('../../models/CommandArg')
+const { Erc20TokenServiceBuilder } = require('../../factories')
 
 /**
  * Class representing GetAllowanceCommand. */
