+++ src/factories/commands/CreateWalletCommandBuilder.js
@@ -1,7 +1,7 @@
 const logger = require('../../logger')
 const { PrivateKeyService } = require('../../services/PrivateKeyService')
 
-const CreateWalletCommand = require('../../commands/CreateWalletCommand')
+const CreateWalletCommand = require('../../commands/trading-wallet/CreateWalletCommand')
 const CreateWalletCommandValidator = require('../../validators/CreateWalletCommandValidator')
 const GetAllowanceCommandValidator = require('../../validators/GetAllowanceCommandValidator')
 const PrivateKeyValidator = require('../../validators/PrivateKeyValidator')
