+++ src/factories/commands/DepositEthCommandBuilder.js
@@ -1,7 +1,7 @@
 const logger = require('../../logger')
 const { PrivateKeyService } = require('../../services/PrivateKeyService')
 
-const DepositEthCommand = require('../../commands/DepositEthCommand')
+const DepositEthCommand = require('../../commands/trading-wallet/DepositEthCommand')
 const DepositEthCommandValidator = require('../../validators/DepositEthCommandValidator')
 const PrivateKeyValidator = require('../../validators/PrivateKeyValidator')
 const TradingWalletServiceBuilder = require('../TradingWalletServiceBuilder')
